requirements.txt
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.39
click==6.7
decorator==4.0.11
Django==1.10.6
docutils==0.13.1
elasticsearch==5.2.0
futures==3.0.5
geocoder==1.20.1
jmespath==0.9.2
kafka-python==1.3.3
nltk==3.2.2
oauthlib==2.0.1
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
ratelim==0.1.6
requests==2.13.0
requests-aws4auth==0.9
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
textblob==0.12.0
tweepy==3.5.0
urllib3==1.20
